pyproject.toml
[project]
name = "k8s-thw-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "libvirt-python>=11.7.0",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
